move send to gosnmp.go, fix errorf in marshal.go
